notiGenerater 생성
